ADD: metrics
This adds in the first pass on doing the metrics. They are simple but
work. Also, the duration metric seem to be incorrect. This needs to be
tracked down to what is causing the issue.

Some refactor has being done as it is now starting to come clear where
the lines between the library and application may be.

Signed-off-by: Jim Fitzpatrick <jfitzpat@redhat.com>